refactor: common code for criterion benches
Closes #1354
